+++ src/entrypoint/directory.h
@@ -54,6 +54,11 @@ struct directory {
      */
     coro<std::size_t> data_size();
 
+    /**
+     * Return amount of effective data volume stored in all buckets.
+     */
+    coro<std::size_t> effective_data_size();
+
 private:
     db::database& m_db;